Infra: fix 2.6 lobby nginx config (#11015)
* Infra: fix 2.6 lobby nginx config

(1) Fix mapping statement to match header name

Nginx automatically converts dashes in headers
to underscores. In nginx config, we need to
specify header names with underscores.

(2) Update SSL cert paths as managed by certbot

(3) Use IP address instead of localhost

Nginx uses a resolver when there is a variable in 'proxy_pass'
to resolve the proxy_pass hostname.
See: https://stackoverflow.com/questions/17685674/nginx-proxy-pass-with-remote-addr
So, when we have a variable in 'proxy_pass', we are required
to specify a 'resolver' directive. We do not need a resolver
at all and instead can use the IP address of localhost directly.

* Add override for 'reverse_proxy_server_name', otherwise defaults to localhost