Game-Client: run game-notes migration when loading game notes
Performance improvement. Moves "xml" to "html" game-note
migration from program launch to instead be done (on-demand)
when we read game-notes.

This avoids a possible performance hit when launching the game.
'PerfTimer' has the game notes migration process as needing
about 30ms (so we are adding a pretty minimal one-time time
cost to reading game notes)