Add a game-version label to error reports
This update will have us add a 'version' label to github issues
that are created as error reports. This will be useful to help
filter error reports by game version. As part of this update
we also remove the 'version' value that was previously present
in the title of the error report message (now redundant to the
new version label)