Build checks: remove check of SQL entity names
Removing because possibly triggers on false-positives.
The check for SQL naming validated that we had nothing camelCased
in any of the '*.sql' files. While generally true, this can fail
when being scanned against data in 'insert' statements.

Rather than building in a way to override this to create exceptions,
for now the check is being removed.