Infra: bring bots back to 2.5 & use syslog
2.6 launch was a bust, this update brings all bots back to the 2.5 lobby.

This update also adds configs for bots to send their log output to syslog.
We add some extra config for syslog to direct this log output to the
file system (/var/log/triplea/), so we can now easily view log output
by checking out that folder and the log files within it.