Smoke-testing: remove bot-client-lobby checks & move smoke-testing to game-app
Smoke testing does a number of things, verifies we can load older save games,
that we can parse maps, and that we can launch a lobby instance and connect
to it with a bot & game client.

This update removes the very latter part of launching a lobby. The setup for
that is somewhat complex & we wish to do little more with bots other than
delete them. Further, the testing is not that in depth, so we do not lose
that much (but we do lose some testing..).

After removing the lobby launch smoke-test, all remaining parts of the
smoke testing are related to the game engine.