Fix stalemate logic to end 0 att plane vs. transport fights. (#11736)
Without this, we would trigger the 10K rounds limit in AI simulation when the battle ends up with a strat bomber (that has no attack or defense) faced off against a transport and/or subs, and neither would be able to damage the other.

Added a unit test. Also renames a couple of existing tests for clarity.

Fixes: #10936

Note: This problem does not exist in 2.5 (tested).