Fix duplicate key error with edit mode's unit deletion. (#11747)
Fix duplicate key error with edit mode's unit deletion.

This can happen if a unit chooser is brought up and the user manually selects both a transport and a unit loaded onto it. The unit loaded onto it will be added twice due to also being added as a dependent.