Omit attachment order data from saves by default. (#11771)
* Omit attachment order data from saves by default.

This data is only needed for the "export to XML" function which is only used by mapmakers. However, the data is a huge contributor to save file size (and saving time). With this change, Imperialism 1974 saves go from 1.2MB in size to 800K (1/3 reduction).

The functionality to export to XML is maintained by re-reading the attachment order data from the original XML file when the export function is used.

Tested exporting map XML and confirmed it still worked.