Make CollectionUtils guarantee mutable return value.
Previously, they used Collectors.toList() which did not provide any such guarantee in its API docs, but behind-the-scenes was implemented as returning a mutable collection.

As such, there's likely triplea code that depends on that implementation detail. This change makes that implementation detail an API guarantee.

Also cleans up a couple code places that were copying the result and a few other clean ups.