Fix "Bombing Raids should be dealt with first!" error.
This could happen if a unit is sent to bomb somewhere but then gets moved again, resulting in an empty battle (which gets skipped during the bombing battle phase).
This removes the logic for considering empty for the purposes of this error. Also removes code to cancel such battle right before throwing the error, since the expectation is that we shouldn't be hitting this codepath, so no need to try to "handle" the situation.

Also, some related cleanup.