PlacePanel: Limit max based on unit requirements.
This change caps the max for each unit type in the place panel to the max that can be placed for that type when considering "units which require units" logic (i.e. factory types).

This partially prevents the error "Cannot place more units which require units, than production capacity of territories with the required units" after making the selection in the place panel.

Note: That error can still happen when selecting multiple units of different types that surpass the max.