Clear transportedBy on fighters on allied carriers when battle doesn't happen. (#12041)
What was happening was that transportedBy was being set when the battle was created, but then if the units moved out still during combat move, it was not being cleared. This caused those fighters later on to have incorrect state and to not be able to participate in future combats.

Fixes: #10294
(Although the existing save will not be fixed, it will just make the issue not happen in new games.)

This is a pre-2.6 bug.

Includes a test and some related code/comment clean ups.