Build: split tests that need database from those that do not
Update adds a tag to tests that need a database. We then
split the build such that 'test' no longer runs these
tests, and 'testWithDatabase' does (and we add
'testAll' to test all)

This allows for './gradlew test' to no longer depend
on any running docker containers & faster game-headed
development.