Build: Remove 'check' task when building game-installers (#12129)
We already run './gradlew check' against merges to master
and any PR. Thus, the check as pre-requisite for building
game installers is redundant.

This update opens up a possibility that a 'check' could
fail after merging to master & we still build game
installer artifacts. In this case we will still get
notification that the main branch build has failed.
Such a situation should be rare enough that the simplicity
in not having to configure our branch tests twice, is
worth the trade-off.