Automatically format when 'verify' script is run
This update changes the 'verify' script to also
run 'spotlessApply'. This automatically fixes any
formatting issues.

Meanwhile, we also update the CI build to skip the
'spotlessApply' task. This means if a developer
runs 'verify', they will automatically get formatting
fixes; but if a branch with formatting problems is pushed
it would fail the build.

For context, the `verify` script is run by CI to run
all build verifications (AKA: checks, tests). This
same script can be run by a developer as-is on their
system to test their local branch.