Build: Drop custom check for unused DB-unit test files (#12143)
The custom check is nice, but.. it is custom. There is
a dependency on paths in the script that make it a bit
brittle. If we widen up the script to search all paths,
for db-test-files as well as the tests that use them,
then we run into performance problems.

All build checks need to have really, really high value
and good performance. Thus, instead of repaing this build
check, just dropping it instead.