Simplify: merge class 'LobbyClient' into 'LoginResult'
This update replaces any usage of LobbyClient with LoginResult.
There is a lot of overlap between the two classes, for the most
part it is a 1:1 replacement.

LobbyClient itself has some issues, it's certainly more of a
result object and has a lot of overlap with another result object
named 'LoginResult'.