Remove unnecessary BidMode.NOT_BID check.
This check is not needed, because the BID codepath this is trying to handle is impossible now.

For it to be possible, maxPlaceableMap would need to have an entry with 0 in bid mode. But this can't happen because:
  - producers will always just have a single territory of at due to the override impl for bid
  - maxPlaceableMap will never have an 0-entry, since it calls getMaxUnitsToBePlacedFrom() to fill the map and that function is overridden for bid to return -1 or units.size()

So this should have no functional changes and all the tests pass.