Image fix (#12507)
Fixes: #12500

Issue is difference between: getImage(imageKey) vs getTransformedImage(imageKey)
getImage returns a scaled image, getTransformedImage returns an unscaled image.

Issue was introduced in: 1ab27d4, there was a mistake to swap 'getTransformedImage' with 'getImage'. Without the clarification that one returns a scaled image and the other does not, it really does look they ought to return the same result. The multiple layers of caching also makes this confusing. It's very surprising for the 'image icon' cache to store unscaled images while the 'images' cache stores scaled images (one would expect both of them to store scaled images).

The problem with the purchase screen is that the scaled images were being used, rather than the 100% scale images.