Leverage existing JAR publish of existing http-client subproject
This update unifies the newly added JAR publish config with the existing
config.

I did not realize we already published jars, hence the need to unify
config that does about the same. Even better, the existing mechanism
only publishes the 'http-client' JAR, which is exactly
(and no more than) what we need for a server dependency.