Fix Bot autosave bug: include bot name in save game file
In autosaves, we want the bot name to be included in the save game file
so that multiple bot instances on the same server do not overwrite each
others files.

Prevoiusly there was a bug "System.clear" of the bot name, thereby
erasing it from the save game file. With that, all bot hosts on the
same server were overwriting each others file. A further consequence,
without the name of the bot instance, it was difficult to know
whose autosave file was whose.

This update removes the unnecessary "System.clear" and with that, the
bot name appears in the autosave file.