Bump version to 2.7 & Bot/Host Updates
(1) Update game selection on staging screen. This introduces incompatibility
with bots, hence game version number is bumped to 2.7. Game selection updates
removes the options to "change to autosave odd" etc in favor of the save
game list available in the game selection list. The former autosave selections
are hopelessly broken since they assumed that the autosave file did not contain
the bot name. That was now fixed, thereby invalidating that autosave functionality.

(2) Favor strict use of system properties instead of CLI args. The old mechanism
parsed CLI args and then set each one in turn as a system property. We can just
use system properties directly to begin with and avoid the parsing & translation
layer.

(3) Add config to run bots with docker

(4) Update README instructions to be current for the variety of ways to run
a bot.