Map extract fix: add replace flag to overwrite bad zips (#12760)
If a bad map is downloaded and can't be extracted, it is moved
into the 'bad zips' folder. If the same map is redownloaded,
then we have a failure moving the bad zip because the target
folder already exists.

This is a patch to replace the bad zip if it already exists.