Startup behavior: Map extraction enhancements (#12766)
Changes behavior of the system when zipped map files are encountered on startup.

For context, current system will:
- see zip folder in downloaded maps folders
- create a temp folder in the downloaded maps folder
- extracts the zip file to the new temp folder
- creates the target map folder and moves all extracted files to that folder
- deletes the zip file

Changes:
- zip file is now extracted to a temporary folder in the systems temporary folder.
This way if the process fails later on, the temp files will not be seen as potential maps
(which can create duplicates). Additionally, any such files will now be automatically
cleaned up, the system temp folder is usually emptied on system reboot.

- smarter attempt to move zip files to a 'bad-zip' folder. Only attempt to do so if
the 'bad-zip' still exists

- cleanup temporary folder where map files are extracted. The folder was not being deleted
in a successful extraction case.