Fix custom check script to return correct status code
The 'parallel' call is having a side effect where if we set
"status=1", the variable assignment is lost after 'parallel'
returns. So, if a check fails, we return "OK" incorrectly
because the statuc code is reverted.

This update removes the 'parallel' (which has the drawback
of a slower script), but corrects the output to correctly
flag errors