Github action packaging: INSTALL4J license variable potential fix
Set env variables a different way. It appears our license variable is
not being exported and is therefore not available to the 'gradle'
child process that is started via our 'package' script.

This update does a github action config to set env variables that
are (hopefully) exported and therefore (hopefully) available
to the gradle process that needs the INSTALL4J_LICENSE_KEY variable.