Dev tools - docker components & scripts to run a lobby server (#13476)
Add a number of docker configs & scripts to a new folder '.docker'

We add docker-compose configuration that can launch a lobby on localhost
port 5000. We add convenience scripts to start & stop the lobby.

Soon we'll have more servers and more databases to start up and not just
the lobby. To handle this we'll want to have a NGINX do routing for us,
which means adding a docker container with NGINX and the correct
routing config for localhost.