Bulk fix triplea_maps.yaml image URLs
(1) The URLs of the form 'github.com..../raw/...' redirect to 'raw.githubusercontent.com'
These URLS are replaced with their redirection targets. Curling these URLs now returns a 200 rather
than a 302. When validating image URLs we can now exclude 302s as not a valid status code.

(2) The URLs of the form 'github.com..../blob/...' pointed to the HTML page containing
the preview image. These URLs return HTML content, not an image! These are corrected.