Remove IllegalState checks (#13961)
The checks look correct, but also when they fire, the game
looks correct as well. It seems like we have eventual consistency,
and having small errors in history is not a big deal.

Removing the illegal state checks allows the game to play
as AI players for many rounds, without - the game barely
makes it out of round 1