Remove crash when triggers subtract PUs below zero
An arbitrary trigger can subtract a fixed PU amount.
This might be more than the current number of PUs.
When adding such an event to the history, there was a
check that crashed the game if PUs would go below zero.

This update changes that logic to allow this update,
so you can subtract an arbitrary amount of PUs.
Except, PUs cannot go negative, we stay at zero
instead. A future update could make this behavior
configurable. For now, this is a fix to the game
crash situation.